[ARM] tegra: enable l2x0 prefetching with 8-cacheline offset
enable data prefetching in the PL310 L2 cache, with an offset of 8
cachelines. this value was empirically derived from memory latency
and memory copy benchmarks running on a single CPU at 1GHz (second
CPU was idle or hot-unplugged), and selected as the value which
provided the most benefit and the least penalty.

memcpy improves from ~380MiB/sec to ~420MiB/sec on Harmony with
this change.

Change-Id: Ied82be59f3fc4735144929a3af2ff0fc30cd970f