[ARM/tegra] nvrm: clean up RM GPIO API implementation
the gpiolib function gpio_direction_output expects to have the value
asserted on the pins provided as a parameter, but the RM GPIO code
was always providing 0; this was causing numerous bootup problems for
user-space GPIO access

on AP20, VI GPIOs can be accessed through the standard GPIO controllers,
so this cleaned-up implementation just reassigns legacy VI GPIOs to the
correct port/pin on the proper GPIO controllers, to eliminate the need
for two codepaths.

additionally, since no one is using the ODM External GPIOs, this change
removes support for them.

and, since the only chip-specific codepath in this tree is just the VI-to-
GPIO mapping, move it into the common directory.

Change-Id: I182d161a05548d299a4a48714b37c2cd57a1e77c