tegra nvos: break single mega node into multiple nodes
Convert /dev/nvos to a general purpose node, /dev/nvos and a super
user node, /dev/knvos. /dev/nvos exposes semaphore related
functionality whereas /dev/knvos will be used for interrupt
and memory management. Track the semaphores in a rbtree
so that we do not stomp on anybody else's space and corrupt
the memory.

Change-Id: I98bacb7312fd3eb86122dde80381b74ae865ce57
Reviewed-on: http://git-master/r/2192
Tested-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>