[tegra ALSA]: Fix hang when CLOSE is delayed after STOP.
If CLOSE call is delayed after STOP, it set's the prtd->state
to SNDRV_PCM_TRIGGER_STOP causing the STOP case execution twice
in play_thread which leads to hang as mixer doesn't respect
and don't send the notification of completion for second mixer
stop done call which is the expected behaviour.

Putting a flag to ensure that stop is executed only once in
play and record thread.

We are setting prtd->state to SNDRV_PCM_TRIGGER_STOP specifically
in CLOSE as application directly call CLOSE without STOP in some
cases.

For Bug: 697124

Change-Id: I0a543bca2488afe0bacbf6497df49f24aecb46bc
Reviewed-on: http://git-master/r/2659
Tested-by: Manjula Gupta <magupta@nvidia.com>
Reviewed-by: Vijay Mali <vmali@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>