tegra: E1206: touch-driver: Enable multi-touch suspend/resume
Driver level changes required to enable multi touch suspend/resume.

Suspend and resume functionality depend on a GPIO connected to
the reset pin of the touch hardware. This is automatically detected
via a odm query and if reset pin is available for the board,
suspend and resume are enabled for that board. If not, suspend and
resume calls will be no op.
During resume process, as soon as we clear the sleep mode bit in
the the hardware, till the time the scan mode is enabled, all the
interrupts received are spurious and since INT is set on resume,
to ignore these spurious interrupts when the resume process starts,
the corresponding GPIO INT is masked and is un-masked only after
spurious INT bit is cleared and scan mode is enabled.

Tested on: E1206

Change-Id: I518b7656057e14533e8c7b7a75bdc3b285f7633f
Reviewed-on: http://git-master/r/2568
Tested-by: Jitendra Aditya Lanka <jlanka@nvidia.com>
Tested-by: Hanumanth Venkateswa Moganty <vmoganty@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>