[arm/tegra] serial: Removing tx fifo empty wait after each tx.
When dma is used to transmit the data through uart, there is a wait
to become the tx fifo empty after each tx dma complete.
This wait is not required after each dma transfer complete because
the next dma/pio based transfer can start even if tx fifo is not
empty. Th uart controller only req dma when there is space in tx
fifo.

The correct state will be returned from the function of tx_empty
based on tx state and fifo status.

Change-Id: I3eb34e26d68b8a3d214759c23b89ef6585881729
Reviewed-on: http://git-master.nvidia.com/r/3782
Reviewed-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Tested-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>