[kernel/driver]: Driver for the magnetometer AK8975
Adding the new kernel driver for the magnetometer which provides the
compass functionality.

This kernel driver provides the sys fs interface to change the property
and read the flux information. Following sysfs interface are available:
mode: set the opertaing mode as 0 for power down and 1 for conversion.
show_flux: To get the current flux data in x, y, z axis. This is blocking
call and conversion start once this file is read. The return
value in microTesla (uT).

The driver can be enabled by setting the CONFIG_HWMON to y and
CONFIG_SENSORS_AK8975 to y in the defconfig. The board information
for device can be provided through odm init code.

Change-Id: If9fe4873835f65579254669aeb915717eeb30834
Reviewed-on: http://git-master.nvidia.com/r/2544
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>

Conflicts:

	drivers/hwmon/Kconfig