[ARM] tegra: suspend: add work-around for AP20 A03 LP0 restore
SDRAM initialization is broken in the boot ROM on AP20 A03, so LP0 restore
needs to reload the first-stage bootloader and then jump to its LP0
restore sequence.  to trigger this behavior, the lp0 flag is moved to a
dummy location in the scratch register which the boot ROM ignores, but
the bootloader will detect.

Change-Id: Iea144e7440367938755ab66d67758558bc184b44
Reviewed-on: http://git-master/r/4022
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>