dsi: Configuring Normal mode for lcd_de pin
The POR value of the lcd_de pin is pull up. Becasue of
default pull up of lcd_de pin, display controller is
triggering the frame even when TE is not connected to
the panel. So, configuring the lcd_de pin in Normal mode
instead of default pullup. With this change, Android is
display is seen on the dsi panel only when TE is connected
to the panel, else there will not be any display.
Tested on: Whistler, Android using dsi panel.
Bug 703400

Change-Id: I0f934d703caedac1be2317bfe80568d3a74e21db
Reviewed-on: http://git-master/r/4126
Tested-by: Venkata Nageswara Penumarty <vpenumarty@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>