[arm/tegra] Workaround for ARM erratum 742230
Description summary
Under rare circumstances, a DMB instruction between 2 write
operations may not ensure the correct visibility ordering
of the 2 writes.

Implication summary
This erratum is more likely to be exhibited by code taking/releasing
a semaphore. The implication of the erratum is that an external agent
may observe the release of the semaphore before the new payload is
visible, meaning that it can access an out-of-date value for the
payload data.

Workaround summary
A software workaround is available, which consists in setting bit [4]
in the undocumented Diagnostic register. This bit (“Disable DMB-lite”)
causes the DMB to behave as a DSB, ensuring the correct ordering of
the 2 writes.

Requires Cortex-A9 MPCore config with 2 or more CPUs
Present in r1p0 / r1p1 / r1p2 and r2p0 / r2p1 / r2p2

Bug 716466

Change-Id: Idabd534c383d157cdcdbb6e1d79e7bab77e0a36e
Reviewed-on: http://git-master/r/4614
Reviewed-by: Scott Williams <scwilliams@nvidia.com>
Tested-by: Scott Williams <scwilliams@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>