[arm/tegra] serial: write accessor should not add read ops.
The write accessor function uart_writeb() and uart_writel() is
doing the read of the same address on which it is writing the data.
This is causing unnecessarily read of rx fifo and so causing data
loss in rx path if tx fifo is getting written by cpu. This is
happening becasue of rx fifo and tx fifo address are same.

Change-Id: I194363872d0fd251ddd15a40f42e58acd5ccc7a1
Reviewed-on: http://git-master/r/4746
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Venkata (Muni) Anda <vanda@nvidia.com>
Reviewed-by: Ramachandrudu Kandhala <rkandhala@nvidia.com>
Reviewed-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Tested-by: Anantha Idapalapati <aidapalapati@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>