[arm/tegra] kbc: Adding scanning timeout configuration.
Keyboard controller goes to continuous polling mode from
wakeup mode if any key is pressed and stay for hw default
to 5 second. This parameter was not configurable from odm.
It is require to configure this parameter from odm based on
platform.
Adding support for keeping this time as configurable. The default
time still be 5 second if it is not configured from odm.

Change-Id: I0fc11cc2a2d64db33b00c048f87081b6181354fa
Reviewed-on: http://git-master/r/8644
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Chun-Hung (Alex) Wu <chuwu@nvidia.com>
Tested-by: Chun-Hung (Alex) Wu <chuwu@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>