[arm/tegra]dma: protecting the dma transfer count update
When client asked for the dma transfer count, then the dma
status register read and update is protected by the spin lock
so that it should not be updated incorrectly.
If client asked for transfer count between dma interrupt occurs
and dma isr thread executed and repeat mode is selected then
updating the transfer count at the time of calculating transfer
count.

bug 732197
bug 731403

Change-Id: Id112d5bea9c2f4a1fc4d329d5ce04eed477f1dfa
Reviewed-on: http://git-master/r/11019
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Udaykumar Rameshchan Raval <uraval@nvidia.com>
Tested-by: Pradeep Goudagunta <pgoudagunta@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>