yaffs patch-ker.sh doesn't work with Linux 3.0-rc1. get_sb field removed from file_system_type in 2.6.39
The test for 2.6.x in patch-ker.sh fails with kernels 3 and up.

A fix is trivial but as I am not profficient with git and am discouraged
from fixing it directly.

Also, 2.6.39 has removed the .get_sb field and only supports the .mount
field in file_system_type struct. I attach a non-git patch that seems to
work ok. It tests for kernel version >= 2.6.39 but as this probably applies
to previous kernels it may be better to #define USE_MOUNT_FIELD conditional
of some previous kernel threshold and use that instead.

Nick Bane