star_battery_charger: Fix refresh rescheduling during deepsleep
While in suspend, the batt driver uses the RTC to schedule the
next round of status updates; the RTC, however, isn't capable
of time-travel, so setting a wake-up for the present or the
past is rejected.
Since the refreshes are recursively set, this completely stalled
the entire update cycle, and stopped any future RTC alarms from
happening.

So... try to make sure any scheduling uses a future date.